+++ arch/arm64/kvm/arm.c
@@ -39,9 +39,8 @@
 #include <asm/kvm_nested.h>
 #include <asm/kvm_pkvm.h>
 #include <asm/kvm_emulate.h>
+#include <asm/kvm_pmu.h>
 #include <asm/sections.h>
-
-#include <kvm/arm_pmu.h>
 
 #include "arm_hypercalls.h"
 #include "arm_psci.h"
